test: rename `strict` to `normal`
