+++ testdata/rule/start_with_name/normal/good_pkg_test.go
@@ -1,4 +1,4 @@
-package strict
+package normal_test
 
 // (BG: bad godoc)
