+++ testdata/rule/start_with_name/normal/good_test.go
@@ -1,4 +1,4 @@
-package strict_test
+package normal
 
 // (BG: bad godoc)
