+++ testdata/rule/start_with_name/normal/good_with_article.go
