+++ testdata/rule/start_with_name/normal/missing.go
@@ -1,4 +1,4 @@
-package strict
+package normal
 
 // (BG: bad godoc)
